requirements.txt
pinecone
sentence-transformers
dotenv
google-genai
python-dotenv
PyPDF2
fastapi
pydantic
asyncpg
